+++ src/fs/directory.cc
@@ -26,7 +26,7 @@ namespace eclipse {
     auto file = reinterpret_cast<FileInfo*>(file_info);
     file->name          = argv[i++];
     file->hash_key      = atoi(argv[i++]);
-    file->size          = atoi(argv[i++]);
+    file->size          = atoll(argv[i++]);
     file->num_block     = atoi(argv[i++]);
     file->type          = atoi(argv[i++]);
     file->replica       = atoi(argv[i]);
